requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-dotenv
motor
apscheduler
kiteconnect
tavily-python
